refactor(folder): group files into feature folders
