+++ Backend/src/ShoeStore.Api/Controllers/ProductEmbeddingController.cs
@@ -1,7 +1,7 @@
 using Asp.Versioning;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using ShoeStore.Application.Interface;
+using ShoeStore.Application.Interface.ChatBotInterface;
 
 namespace ShoeStore.Api.Controllers;
