+++ Backend/src/ShoeStore.Application/Interface/ChatBotInterface/IProductEmbeddingRepository.cs
@@ -1,7 +1,7 @@
 using ShoeStore.Application.Interface.Common;
 using ShoeStore.Domain.Entities.Embedding;
 
-namespace ShoeStore.Application.Interface;
+namespace ShoeStore.Application.Interface.ChatBotInterface;
 
 public interface IProductEmbeddingRepository : IGenericRepository<ProductEmbedding, int>
 {
