+++ Backend/src/ShoeStore.Application/Interface/CheckoutInterface/IPaymentTransactionRepository.cs
@@ -1,7 +1,7 @@
 using ShoeStore.Application.Interface.Common;
 using ShoeStore.Domain.Entities;
 
-namespace ShoeStore.Application.Interface;
+namespace ShoeStore.Application.Interface.CheckoutInterface;
 
 public interface IPaymentTransactionRepository : IGenericRepository<PaymentTransaction, int>
 {
