--- Backend/src/ShoeStore.Application/Interface/IProductEmbeddingService.cs
@@ -1,10 +1,0 @@
-using ErrorOr;
-
-namespace ShoeStore.Application.Interface;
-
-public interface IProductEmbeddingService
-{
-    Task<ErrorOr<Created>> GenerateVectorEmbeddingWithExistDataAsync(CancellationToken cancellationToken);
-
-    Task<ErrorOr<Created>> GenerateVectorEmbeddingByProductPublicId(Guid productId, CancellationToken cancellationToken);
-}
