+++ Backend/src/ShoeStore.Application/Services/ProductEmbeddingService.cs
@@ -2,7 +2,7 @@ using System.Text;
 using ErrorOr;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.AI;
-using ShoeStore.Application.Interface;
+using ShoeStore.Application.Interface.ChatBotInterface;
 using ShoeStore.Application.Interface.Common;
 using ShoeStore.Application.Interface.ProductInterface;
 using ShoeStore.Domain.Entities;
@@ -11,16 +11,16 @@ using ShoeStore.Domain.Entities.Embedding;
 namespace ShoeStore.Application.Services;
 
 /// <summary>
-/// Generates and persists vector embeddings for products to support semantic search.
+///     Generates and persists vector embeddings for products to support semantic search.
 /// </summary>
 public class ProductEmbeddingService(
     IProductRepository productRepository,
     IProductEmbeddingRepository productEmbeddingRepository,
     IUnitOfWork unitOfWork,
-    IEmbeddingGenerator<string,Embedding<float>> embeddingGenerator) : IProductEmbeddingService
+    IEmbeddingGenerator<string, Embedding<float>> embeddingGenerator) : IProductEmbeddingService
 {
     /// <summary>
-    /// Generates embeddings for all existing products in batches and stores them in the database.
+    ///     Generates embeddings for all existing products in batches and stores them in the database.
     /// </summary>
     /// <param name="cancellationToken">Token to cancel the operation.</param>
     public async Task<ErrorOr<Created>> GenerateVectorEmbeddingWithExistDataAsync(CancellationToken cancellationToken)
@@ -41,6 +41,7 @@ public class ProductEmbeddingService(
                 var textChunk = GenerateText(product);
                 textChunks.Add(textChunk);
             }
+
             var embeddingVector = await GenerateVectorEmbeddingAsync(textChunks, cancellationToken);
             for (var index = 0; index < productList.Count; index++)
             {
@@ -52,22 +53,25 @@ public class ProductEmbeddingService(
                 };
                 productEmbeddingList.Add(newProductEmbedding);
             }
+
             productEmbeddingRepository.AddRange(productEmbeddingList);
             await unitOfWork.SaveChangesAsync(cancellationToken);
         }
+
         return Result.Created;
     }
 
     /// <summary>
-    /// Generates and stores the embedding for a single product identified by its public ID.
+    ///     Generates and stores the embedding for a single product identified by its public ID.
     /// </summary>
     /// <param name="productId">Public product ID used to locate the product.</param>
     /// <param name="cancellationToken">Token to cancel the operation.</param>
     /// <returns>Created result on success or an error when the product is not found.</returns>
-    public async Task<ErrorOr<Created>> GenerateVectorEmbeddingByProductPublicId(Guid productId, CancellationToken cancellationToken)
+    public async Task<ErrorOr<Created>> GenerateVectorEmbeddingByProductPublicId(Guid productId,
+        CancellationToken cancellationToken)
     {
         var product = await productRepository.GetDetailsByGuidAsync(productId, cancellationToken);
-        if(product is null) return Error.NotFound("Product.NotFound","Product not found");
+        if (product is null) return Error.NotFound("Product.NotFound", "Product not found");
         var textChunk = GenerateText(product);
         var textChunks = new List<string> { textChunk };
         var vectorEmbedding = await GenerateVectorEmbeddingAsync(textChunks, cancellationToken);
@@ -83,7 +87,7 @@ public class ProductEmbeddingService(
     }
 
     /// <summary>
-    /// Builds a text representation of a product for embedding generation.
+    ///     Builds a text representation of a product for embedding generation.
     /// </summary>
     /// <param name="product">Product with variants, category, and brand.</param>
     /// <returns>Flattened text describing the product and its variants.</returns>
@@ -106,12 +110,13 @@ public class ProductEmbeddingService(
     }
 
     /// <summary>
-    /// Generates vector embeddings for the provided text chunks.
+    ///     Generates vector embeddings for the provided text chunks.
     /// </summary>
     /// <param name="textChunk">Ordered list of text chunks to embed.</param>
     /// <param name="token">Token to cancel the operation.</param>
     /// <returns>Embedding vectors aligned with the input order.</returns>
-    private async Task<List<ReadOnlyMemory<float>>> GenerateVectorEmbeddingAsync(List<string> textChunk, CancellationToken token)
+    private async Task<List<ReadOnlyMemory<float>>> GenerateVectorEmbeddingAsync(List<string> textChunk,
+        CancellationToken token)
     {
         var vector = await embeddingGenerator.GenerateAsync(textChunk, cancellationToken: token);
         var result = vector.Select(v => v.Vector).ToList();
